Initial commit for Android implementation
The approach we took is fairly simple. Navigation is created natively in corresponding activity,
currently only TabActivity is supported, and we pass the styling parameters and screenId to it.
Each screen is rendered in its own ReactRootView and is wired to the same ReactInstanceManager,
so all screens share the same java script context.