Support defining MenuItem action from JS
Supported actions:

1. never - Never show this item as a button in an Action Bar.

2. always - Always show this item as a button in an Action Bar.
Use sparingly! If too many items are set to always show in the Action
Bar it can crowd the Action Bar and degrade the user experience on
devices with smaller screens. A good rule of thumb is to have no more
than 2 items set to always show at a time.

3. withText - When this item is in the action bar, always show it
with a text label even if it also has an icon specified.

4. ifRoom (default) - Show this item as a button in an Action Bar if
the system decides there is room for it.