Check if tab exists before applying tab options
Child options are applied when a child appears, meaning it's attached to hierarchy.
Going over the Wix app crash logs, a few users experienced a crash where apparently
BottomTabs options were applied for a child component, and a corresponding tab wasn't found.
I'm not sure how this happens and wasn't able to reproduce the crash.

Hopefully I'll have more information in the future but for now I've added a simple check to ensure
options are applied only if a tab is found.