Update SideMenu enabled in parent (#6196)
When a menu is opened its resolved options are applied. These options set the enabled property of both left and right menus. If a menu is disabled, opening the other menu will unintentionally enable it when it is opened therefore we need to update the enabled properties of both menus in the actual SideMenu layout.

Closes #6182

Co-authored-by: Yogev Ben David <yogev132@gmail.com>