Mount all stack children (#6194)
This commit somewhat consolidates how stack children are mounted.

On Android, until now, only the top child in a stack was mounted when the stack became visible or when setStackRoot() was called. When popping the stack the new current child would be mounted when it was attached. This caused flickering when popping the stack as the pop animation started before the child finished mounting.

Now all children will be mounted together after the top child is mounted :)

Co-authored-by: Yogev Ben David <yogev132@gmail.com>