Handle errors in postlink scripts (#6267)
This PR aims to provide some basic error handling logic to the postlink scripts, which run after `react-native link`, by identifying potential drawbacks and managing the execution flow via try/catch statements and the use of throw under specific conditions. It was developed with the support of the @underscopeio team!

I added a few messages I thought are informative, which provide some basic feedback and direct the user to the specific section in the documentation whenever it's possible. The script was mostly optimistic, so I changed this a little bit to throw a few errors when unexpected things occur. On the previous version, errors would either be silent or be incorrectly informed as the library being already linked, as this could not be asserted from the information available during execution.

Also, I replaced a few single quote/double quote discrepancies on the files I worked on (went for double quotes since that was predominant among the scripts). But adding linting/prettier to these scripts would be nice too, this could be done in a future PR if there's interest.

I'd also like to address the issue of the deprecation warning for `react-native link` on a separate PR, I have a candidate solution that should require minimum changes on the docs and the process. I will send this shortly.