Moved website code snippets to own files (#6512)
Fixes #6495 

This allows website code snippets to be build with `tsc` reducing their error proneness.

I converted all code snippets tagged with `tsx?` to separate files. There are other snippets tagged as `jsx?` which I did not convert (yet, at least out of scope for this pr).

The chosen directory structure represents the ability to colocate snippets with their respective `.mdx` file. However, we can't move the `.mdx` files into the folder and name them `index.mdx` because that is not supported by docusaurus yet. See https://github.com/facebook/docusaurus/issues/2537

The last commit makes changes to the snippets to make them build. I did not modify the mobx snippets (so they error out at the moment) as I have never used mobx. Somebody else needs to take care of them.

The build can be triggered with the npm / yarn task `snippets:build` in the `website` directory. Later on we should add this to the build pipeline so that errors out if the snippets / rnn lib code are out of sync.