Fix waitForRender + mergeOptions not working when view controller is in transit (#6702)
When `waitForRender` is set to `true`, calling `Navigation.mergeOptions` early (for example in `componentDidMount`) will lead to the options not being applied as the backing view controller has not yet been placed in the window hierarchy. This PR aims to keep track of any pending view controllers and opens them up to being found via their component id. Thus enabling `Navigation.mergeOptions` to work on view controllers in transit.

The pending view controllers are now being kept in `RNNLayoutManager`. For that reason, the layout manager was converted to be instantiable.

A `NSHashTable` was chosen as the backing storage to support keeping the view controllers as a weak reference.

Fixes #6085