Fix ReactImageView sporadic flicker in pop during Shared Element Transition (#6708)
When animating shared elements in pop, the appearing screen needs a bit of time to settle in and for the React images to be displayed.
Seems like a good old post does the job in this case. This optimization is done only for ReactImages. For FastImage, we continue checking their measured bounds as before.