Hide log box in TopBar components (#7062)
Up until RN 62, YellowBox was used to show warnings. RNN was able to detect those and show them only on components used as screens.
RN 63 changed the default implementation to LogBox. With the switch to LogBox we were no longer able to hide warnings on TopBar components.

This commit hides LogBox in topBar components.