Fix broken shared element transition in RTL layouts + fix a flicker when transition started to/from a modal (#7128)
When showing a modal with shared elements, there was a noticeable flicker when the animation started. This happened because the correct alpha wasn't applied to the entering view when the animation was being prepared.

In addition, this commit fixes an issue in RTL. Elements involved in the transition were not positioned correctly. This is resolved by removing the views gravity during the transition.